[lab3] working on duplicate ops

+++ src/kvraft/common.go
@@ -15,7 +15,7 @@ type PutAppendArgs struct {
 	// You'll have to add definitions here.
 	// Field names must start with capital letters,
 	// otherwise RPC will break.
-	ClientId    int
+	ClientId    int64
 	OpSerialNum int
 }
 
@@ -27,7 +27,7 @@ type PutAppendReply struct {
 type GetArgs struct {
 	Key string
 	// You'll have to add definitions here.
-	ClientId    int
+	ClientId    int64
 	OpSerialNum int
 }
